requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
xgboost==2.0.3
lightgbm==4.1.0
joblib==1.3.2
pyyaml==6.0.1 